Updated description for 'actions" and examples
Actions is truly not "actions" but the "Action Groups" value per the deployment code. Recommend changing the description in the readme but leave the param name the same for the sake of existing code that may reference it currently.  Also updated the examples for BICEP and JSON to now include the 'actions' parameter as it was not previously provided and/or left confusion on whether it was the full 'actions' property or just the sub-component of it. 

i.e. - actions parameter input being.....
{
      actionGroups: "/subscriptions/<subsriptionId>/resourceGroups/<resourceGroupName>/providers/microsoft.insights/actionGroups/<actionGroupName>"
      customProperties: {}
}

VS the expected value of just the actionGroups property....

actions: ["/subscriptions/<subsriptionId>/resourceGroups/<resourceGroupName>/providers/microsoft.insights/actionGroups/<actionGroupName>"]